Move layout types to a dedicated types file
These types are only used in the layout classes, so keep them located closed in the folder where they are used.
